# library/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "图书管理系统",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/user",
    "paths": {
...
        "/admin/add": {
            "get": {
                "description": "管理员通过书名、索引号、作者添加书本",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "page"
                ],
                "summary": "管理员添加书籍",
                "responses": {
                    "200": {
                        "description": "获取添加页面成功",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "管理员通过书名、索引号、作者进行添加",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "管理员添加书籍",
                "parameters": [
                    {
                        "description": "书名",
                        "name": "bookName",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "作者",
                        "name": "author",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "索引号",
                        "name": "Pid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "添加成功返回json消息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "添加失败返回失败消息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/delete": {
            "get": {
                "description": "管理员通过书名删除书本",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "page"
                ],
                "summary": "管理员删除书籍",
                "responses": {
                    "200": {
                        "description": "获取删除页面成功",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "管理员通过书名进行删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "管理员删除书籍",
                "parameters": [
                    {
                        "description": "书名",
                        "name": "bookName",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功返回json消息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "删除失败返回json消息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/insert": {
            "get": {
                "description": "管理员通过书名查询书本",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "page"
                ],
                "summary": "管理员查询书",
                "responses": {
                    "200": {
                        "description": "获取查询页面成功",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "管理员通过书名进行查询",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "text/html"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "管理员查询",
                "parameters": [
                    {
                        "description": "书名",
                        "name": "booName",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功返回json消息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "查询失败返回json消息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/login": {
            "get": {
                "description": "管理员获取登录页面",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "page"
                ],
                "summary": "管理员登录",
                "responses": {
                    "200": {
                        "description": "登录页面",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "管理员通过用户名和密码进行登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "管理员登录",
                "parameters": [
                    {
                        "description": "用户名",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "密码",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登陆成功返回json消息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "登录失败返回json消息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/register": {
            "get": {
                "description": "管理员获取注册页面",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "page"
                ],
                "summary": "管理员注册",
                "responses": {
                    "200": {
                        "description": "获取注册页面",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "管理员通过用户名和密码进行注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "管理员注册",
                "parameters": [
                    {
                        "description": "用户名",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "密码",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册成功返回Json消息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "注册失败返回json消息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/update": {
            "get": {
                "description": "管理员通过书名更新书本",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "page"
                ],
                "summary": "管理员更新书籍",
                "responses": {
                    "200": {
                        "description": "获取更新页面成功",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "管理员通过书名进行更新",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "text/html"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "管理员更新书籍信息",
                "parameters": [
                    {
                        "description": "旧书名",
                        "name": "BbookName",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "新书名",
                        "name": "AbookName",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "作者",
                        "name": "Author",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "索引号",
                        "name": "Pid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功返回书本的信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "更新失败返回失败的消息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    }
}